Split web.py
