+++ aiohttp/web.py
@@ -1,30 +1,64 @@
 import abc
 import asyncio
-import binascii
 import collections
 import mimetypes
-import cgi
-import http.cookies
-import io
-import json
 import re
 import os
-import warnings
 
-from urllib.parse import urlsplit, parse_qsl, urlencode, unquote
-from types import MappingProxyType
+from urllib.parse import urlencode
 
 from . import hdrs
 from .abc import AbstractRouter, AbstractMatchInfo
-from .helpers import reify
 from .log import web_logger
-from .multidict import (CIMultiDictProxy,
-                        CIMultiDict,
-                        MultiDictProxy,
-                        MultiDict)
-from .protocol import Response as ResponseImpl, HttpVersion, HttpVersion11
+from .protocol import HttpVersion, HttpVersion11
 from .server import ServerHttpProtocol
-from .streams import EOF_MARKER
+
+from .web_exceptions import (HTTPException,
+                             HTTPError,
+                             HTTPRedirection,
+                             HTTPSuccessful,
+                             HTTPOk,
+                             HTTPCreated,
+                             HTTPAccepted,
+                             HTTPNonAuthoritativeInformation,
+                             HTTPNoContent,
+                             HTTPResetContent,
+                             HTTPPartialContent,
+                             HTTPMultipleChoices,
+                             HTTPMovedPermanently,
+                             HTTPFound,
+                             HTTPSeeOther,
+                             HTTPNotModified,
+                             HTTPUseProxy,
+                             HTTPTemporaryRedirect,
+                             HTTPClientError,
+                             HTTPBadRequest,
+                             HTTPUnauthorized,
+                             HTTPPaymentRequired,
+                             HTTPForbidden,
+                             HTTPNotFound,
+                             HTTPMethodNotAllowed,
+                             HTTPNotAcceptable,
+                             HTTPProxyAuthenticationRequired,
+                             HTTPRequestTimeout,
+                             HTTPConflict,
+                             HTTPGone,
+                             HTTPLengthRequired,
+                             HTTPPreconditionFailed,
+                             HTTPRequestEntityTooLarge,
+                             HTTPRequestURITooLong,
+                             HTTPUnsupportedMediaType,
+                             HTTPRequestRangeNotSatisfiable,
+                             HTTPExpectationFailed,
+                             HTTPServerError,
+                             HTTPInternalServerError,
+                             HTTPNotImplemented,
+                             HTTPBadGateway,
+                             HTTPServiceUnavailable,
+                             HTTPGatewayTimeout,
+                             HTTPVersionNotSupported)
+
+from .web_reqrep import Request, StreamResponse, Response
 from .websocket import do_handshake, MSG_BINARY, MSG_CLOSE, MSG_PING, MSG_TEXT
 from .errors import HttpProcessingError, WSClientDisconnectedError
 
@@ -88,643 +122,9 @@ __all__ = [
 ]
 
 
-sentinel = object()
-retype = type(re.compile('@'))
-
-
-class HeadersMixin:
-
-    _content_type = None
-    _content_dict = None
-    _stored_content_type = sentinel
-
-    def _parse_content_type(self, raw):
-        self._stored_content_type = raw
-        if raw is None:
-            # default value according to RFC 2616
-            self._content_type = 'application/octet-stream'
-            self._content_dict = {}
-        else:
-            self._content_type, self._content_dict = cgi.parse_header(raw)
-
-    @property
-    def content_type(self, _CONTENT_TYPE=hdrs.CONTENT_TYPE):
-        """The value of content part for Content-Type HTTP header."""
-        raw = self.headers.get(_CONTENT_TYPE)
-        if self._stored_content_type != raw:
-            self._parse_content_type(raw)
-        return self._content_type
-
-    @property
-    def charset(self, _CONTENT_TYPE=hdrs.CONTENT_TYPE):
-        """The value of charset part for Content-Type HTTP header."""
-        raw = self.headers.get(_CONTENT_TYPE)
-        if self._stored_content_type != raw:
-            self._parse_content_type(raw)
-        return self._content_dict.get('charset')
-
-    @property
-    def content_length(self, _CONTENT_LENGTH=hdrs.CONTENT_LENGTH):
-        """The value of Content-Length HTTP header."""
-        l = self.headers.get(_CONTENT_LENGTH)
-        if l is None:
-            return None
-        else:
-            return int(l)
-
-
-FileField = collections.namedtuple('Field', 'name filename file content_type')
-
-
 ############################################################
-# HTTP Request
+# Server WebSocket
 ############################################################
-
-
-class Request(dict, HeadersMixin):
-
-    def __init__(self, app, message, payload, transport, reader, writer, *,
-                 _HOST=hdrs.HOST):
-        self._app = app
-        self._version = message.version
-        self._transport = transport
-        self._reader = reader
-        self._writer = writer
-        self._method = message.method
-        self._host = message.headers.get(_HOST)
-        self._path_qs = message.path
-        res = urlsplit(message.path)
-        self._path = unquote(res.path)
-        self._query_string = res.query
-        self._post = None
-        self._post_files_cache = None
-        self._headers = CIMultiDictProxy(message.headers)
-
-        if self._version < HttpVersion11:
-            self._keep_alive = False
-        elif message.should_close:
-            self._keep_alive = False
-        else:
-            self._keep_alive = True
-
-        # matchdict, route_name, handler
-        # or information about traversal lookup
-        self._match_info = None  # initialized after route resolving
-
-        self._payload = payload
-        self._cookies = None
-
-        self._read_bytes = None
-
-    @property
-    def method(self):
-        """Read only property for getting HTTP method.
-
-        The value is upper-cased str like 'GET', 'POST', 'PUT' etc.
-        """
-        return self._method
-
-    @property
-    def version(self):
-        """Read only property for getting HTTP version of request.
-
-        Returns aiohttp.protocol.HttpVersion instance.
-        """
-        return self._version
-
-    @property
-    def host(self):
-        """Read only property for getting *HOST* header of request.
-
-        Returns str or None if HTTP request has no HOST header.
-        """
-        return self._host
-
-    @property
-    def path_qs(self):
-        """The URL including PATH_INFO and the query string.
-
-        E.g, /app/blog?id=10
-        """
-        return self._path_qs
-
-    @property
-    def path(self):
-        """The URL including *PATH INFO* without the host or scheme.
-
-        E.g., ``/app/blog``
-        """
-        return self._path
-
-    @property
-    def query_string(self):
-        """The query string in the URL.
-
-        E.g., id=10
-        """
-        return self._query_string
-
-    @reify
-    def GET(self):
-        """A multidict with all the variables in the query string.
-
-        Lazy property.
-        """
-        return MultiDictProxy(MultiDict(parse_qsl(self._query_string)))
-
-    @reify
-    def POST(self):
-        """A multidict with all the variables in the POST parameters.
-
-        post() methods has to be called before using this attribute.
-        """
-        if self._post is None:
-            raise RuntimeError("POST is not available before post()")
-        return self._post
-
-    @property
-    def headers(self):
-        """A case-insensitive multidict with all headers.
-
-        Lazy property.
-        """
-        return self._headers
-
-    @property
-    def keep_alive(self):
-        """Is keepalive enabled by client?"""
-        return self._keep_alive
-
-    @property
-    def match_info(self):
-        """Result of route resolving."""
-        return self._match_info
-
-    @property
-    def app(self):
-        """Application instance."""
-        return self._app
-
-    @property
-    def transport(self):
-        """Transport used for request processing."""
-        return self._transport
-
-    @property
-    def cookies(self):
-        """Return request cookies.
-
-        A read-only dictionary-like object.
-        """
-        if self._cookies is None:
-            raw = self.headers.get(hdrs.COOKIE, '')
-            parsed = http.cookies.SimpleCookie(raw)
-            self._cookies = MappingProxyType(
-                {key: val.value for key, val in parsed.items()})
-        return self._cookies
-
-    @property
-    def payload(self):
-        """Return raw payload stream."""
-        warnings.warn('use Request.content instead', DeprecationWarning)
-        return self._payload
-
-    @property
-    def content(self):
-        """Return raw payload stream."""
-        return self._payload
-
-    @asyncio.coroutine
-    def release(self):
-        """Release request.
-
-        Eat unread part of HTTP BODY if present.
-        """
-        chunk = yield from self._payload.readany()
-        while chunk is not EOF_MARKER or chunk:
-            chunk = yield from self._payload.readany()
-
-    @asyncio.coroutine
-    def read(self):
-        """Read request body if present.
-
-        Returns bytes object with full request content.
-        """
-        if self._read_bytes is None:
-            body = bytearray()
-            while True:
-                chunk = yield from self._payload.readany()
-                body.extend(chunk)
-                if chunk is EOF_MARKER:
-                    break
-            self._read_bytes = bytes(body)
-        return self._read_bytes
-
-    @asyncio.coroutine
-    def text(self):
-        """Return BODY as text using encoding from .charset."""
-        bytes_body = yield from self.read()
-        encoding = self.charset or 'utf-8'
-        return bytes_body.decode(encoding)
-
-    @asyncio.coroutine
-    def json(self, *, loader=json.loads):
-        """Return BODY as JSON."""
-        body = yield from self.text()
-        return loader(body)
-
-    @asyncio.coroutine
-    def post(self):
-        """Return POST parameters."""
-        if self._post is not None:
-            return self._post
-        if self.method not in ('POST', 'PUT', 'PATCH'):
-            self._post = MultiDictProxy(MultiDict())
-            return self._post
-
-        content_type = self.content_type
-        if (content_type not in ('',
-                                 'application/x-www-form-urlencoded',
-                                 'multipart/form-data')):
-            self._post = MultiDictProxy(MultiDict())
-            return self._post
-
-        body = yield from self.read()
-        content_charset = self.charset or 'utf-8'
-
-        environ = {'REQUEST_METHOD': self.method,
-                   'CONTENT_LENGTH': str(len(body)),
-                   'QUERY_STRING': '',
-                   'CONTENT_TYPE': self.headers.get(hdrs.CONTENT_TYPE)}
-
-        fs = cgi.FieldStorage(fp=io.BytesIO(body),
-                              environ=environ,
-                              keep_blank_values=True,
-                              encoding=content_charset)
-
-        supported_transfer_encoding = {
-            'base64': binascii.a2b_base64,
-            'quoted-printable': binascii.a2b_qp
-        }
-
-        out = MultiDict()
-        for field in fs.list or ():
-            transfer_encoding = field.headers.get(
-                'Content-Transfer-Encoding', None)
-            if field.filename:
-                ff = FileField(field.name,
-                               field.filename,
-                               field.file,  # N.B. file closed error
-                               field.type)
-                if self._post_files_cache is None:
-                    self._post_files_cache = {}
-                self._post_files_cache[field.name] = field
-                out.add(field.name, ff)
-            else:
-                value = field.value
-                if transfer_encoding in supported_transfer_encoding:
-                    # binascii accepts bytes
-                    value = value.encode('utf-8')
-                    value = supported_transfer_encoding[
-                        transfer_encoding](value)
-                out.add(field.name, value)
-
-        self._post = MultiDictProxy(out)
-        return self._post
-
-    def __repr__(self):
-        return "<{} {} {} >".format(self.__class__.__name__,
-                                    self.method, self.path)
-
-
-############################################################
-# HTTP Response classes
-############################################################
-
-
-class StreamResponse(HeadersMixin):
-
-    def __init__(self, *, status=200, reason=None):
-        self._body = None
-        self._keep_alive = None
-        self._chunked = False
-        self._chunk_size = None
-        self._compression = False
-        self._compression_force = False
-        self._headers = CIMultiDict()
-        self._cookies = http.cookies.SimpleCookie()
-        self.set_status(status, reason)
-
-        self._req = None
-        self._resp_impl = None
-        self._eof_sent = False
-
-    def _copy_cookies(self):
-        for cookie in self._cookies.values():
-            value = cookie.output(header='')[1:]
-            self.headers.add(hdrs.SET_COOKIE, value)
-
-    @property
-    def started(self):
-        return self._resp_impl is not None
-
-    @property
-    def status(self):
-        return self._status
-
-    @property
-    def chunked(self):
-        return self._chunked
-
-    @property
-    def compression(self):
-        return self._compression
-
-    @property
-    def reason(self):
-        return self._reason
-
-    def set_status(self, status, reason=None):
-        self._status = int(status)
-        if reason is None:
-            reason = ResponseImpl.calc_reason(status)
-        self._reason = reason
-
-    @property
-    def keep_alive(self):
-        return self._keep_alive
-
-    def force_close(self):
-        self._keep_alive = False
-
-    def enable_chunked_encoding(self, chunk_size=None):
-        """Enables automatic chunked transfer encoding."""
-        self._chunked = True
-        self._chunk_size = chunk_size
-
-    def enable_compression(self, force=False):
-        """Enables response compression with `deflate` encoding."""
-        self._compression = True
-        self._compression_force = force
-
-    @property
-    def headers(self):
-        return self._headers
-
-    @property
-    def cookies(self):
-        return self._cookies
-
-    def set_cookie(self, name, value, *, expires=None,
-                   domain=None, max_age=None, path='/',
-                   secure=None, httponly=None, version=None):
-        """Set or update response cookie.
-
-        Sets new cookie or updates existent with new value.
-        Also updates only those params which are not None.
-        """
-
-        old = self._cookies.get(name)
-        if old is not None and old.coded_value == '':
-            # deleted cookie
-            self._cookies.pop(name, None)
-
-        self._cookies[name] = value
-        c = self._cookies[name]
-        if expires is not None:
-            c['expires'] = expires
-        if domain is not None:
-            c['domain'] = domain
-        if max_age is not None:
-            c['max-age'] = max_age
-        if path is not None:
-            c['path'] = path
-        if secure is not None:
-            c['secure'] = secure
-        if httponly is not None:
-            c['httponly'] = httponly
-        if version is not None:
-            c['version'] = version
-
-    def del_cookie(self, name, *, domain=None, path='/'):
-        """Delete cookie.
-
-        Creates new empty expired cookie.
-        """
-        # TODO: do we need domain/path here?
-        self._cookies.pop(name, None)
-        self.set_cookie(name, '', max_age=0, domain=domain, path=path)
-
-    @property
-    def content_length(self):
-        # Just a placeholder for adding setter
-        return super().content_length
-
-    @content_length.setter
-    def content_length(self, value):
-        if value is not None:
-            value = int(value)
-            # TODO: raise error if chunked enabled
-            self.headers[hdrs.CONTENT_LENGTH] = str(value)
-        elif hdrs.CONTENT_LENGTH in self.headers:
-            del self.headers[hdrs.CONTENT_LENGTH]
-
-    @property
-    def content_type(self):
-        # Just a placeholder for adding setter
-        return super().content_type
-
-    @content_type.setter
-    def content_type(self, value):
-        self.content_type  # read header values if needed
-        self._content_type = str(value)
-        self._generate_content_type_header()
-
-    @property
-    def charset(self):
-        # Just a placeholder for adding setter
-        return super().charset
-
-    @charset.setter
-    def charset(self, value):
-        ctype = self.content_type  # read header values if needed
-        if ctype == 'application/octet-stream':
-            raise RuntimeError("Setting charset for application/octet-stream "
-                               "doesn't make sense, setup content_type first")
-        if value is None:
-            self._content_dict.pop('charset', None)
-        else:
-            self._content_dict['charset'] = str(value).lower()
-        self._generate_content_type_header()
-
-    def _generate_content_type_header(self, CONTENT_TYPE=hdrs.CONTENT_TYPE):
-        params = '; '.join("%s=%s" % i for i in self._content_dict.items())
-        if params:
-            ctype = self._content_type + '; ' + params
-        else:
-            ctype = self._content_type
-        self.headers[CONTENT_TYPE] = ctype
-
-    def _start_pre_check(self, request):
-        if self._resp_impl is not None:
-            if self._req is not request:
-                raise RuntimeError(
-                    'Response has been started with different request.')
-            else:
-                return self._resp_impl
-        else:
-            return None
-
-    def start(self, request):
-        resp_impl = self._start_pre_check(request)
-        if resp_impl is not None:
-            return resp_impl
-
-        self._req = request
-        keep_alive = self._keep_alive
-        if keep_alive is None:
-            keep_alive = request.keep_alive
-        self._keep_alive = keep_alive
-
-        resp_impl = self._resp_impl = ResponseImpl(
-            request._writer,
-            self._status,
-            request.version,
-            not keep_alive,
-            self._reason)
-
-        self._copy_cookies()
-
-        if self._compression:
-            if (self._compression_force or
-                    'deflate' in request.headers.get(
-                        hdrs.ACCEPT_ENCODING, '')):
-                resp_impl.add_compression_filter()
-
-        if self._chunked:
-            resp_impl.enable_chunked_encoding()
-            if self._chunk_size:
-                resp_impl.add_chunking_filter(self._chunk_size)
-
-        headers = self.headers.items()
-        for key, val in headers:
-            resp_impl.add_header(key, val)
-
-        resp_impl.send_headers()
-        return resp_impl
-
-    def write(self, data):
-        if not isinstance(data, (bytes, bytearray, memoryview)):
-            raise TypeError('data argument must be byte-ish (%r)' %
-                            type(data))
-
-        if self._eof_sent:
-            raise RuntimeError("Cannot call write() after write_eof()")
-        if self._resp_impl is None:
-            raise RuntimeError("Cannot call write() before start()")
-
-        if data:
-            return self._resp_impl.write(data)
-        else:
-            return ()
-
-    @asyncio.coroutine
-    def drain(self):
-        if self._resp_impl is None:
-            raise RuntimeError("Response has not been started")
-        yield from self._resp_impl.transport.drain()
-
-    @asyncio.coroutine
-    def write_eof(self):
-        if self._eof_sent:
-            return
-        if self._resp_impl is None:
-            raise RuntimeError("Response has not been started")
-
-        yield from self._resp_impl.write_eof()
-        self._eof_sent = True
-
-    def __repr__(self):
-        if self.started:
-            info = "{} {} ".format(self._req.method, self._req.path)
-        else:
-            info = "not started"
-        return "<{} {} {}>".format(self.__class__.__name__,
-                                   self.reason, info)
-
-
-class Response(StreamResponse):
-
-    def __init__(self, *, body=None, status=200,
-                 reason=None, headers=None,
-                 text=None, content_type=None):
-        super().__init__(status=status, reason=reason)
-
-        if headers is not None:
-            self.headers.extend(headers)
-
-        if body is not None and text is not None:
-            raise ValueError("body and text are not allowed together.")
-
-        if text is not None:
-            if hdrs.CONTENT_TYPE not in self.headers:
-                # fast path for filling headers
-                if not isinstance(text, str):
-                    raise TypeError('text argument must be str (%r)' %
-                                    type(text))
-                if content_type is None:
-                    content_type = 'text/plain'
-                self.headers[hdrs.CONTENT_TYPE] = (
-                    content_type + '; charset=utf-8')
-                self._content_type = content_type
-                self._content_dict = {'charset': 'utf-8'}
-                self.body = text.encode('utf-8')
-            else:
-                self.text = text
-        else:
-            if content_type:
-                self.content_type = content_type
-            if body is not None:
-                self.body = body
-            else:
-                self.body = None
-
-    @property
-    def body(self):
-        return self._body
-
-    @body.setter
-    def body(self, body):
-        if body is not None and not isinstance(body, bytes):
-            raise TypeError('body argument must be bytes (%r)' % type(body))
-        self._body = body
-        if body is not None:
-            self.content_length = len(body)
-        else:
-            self.content_length = 0
-
-    @property
-    def text(self):
-        return self._body.decode(self.charset or 'utf-8')
-
-    @text.setter
-    def text(self, text):
-        if text is not None and not isinstance(text, str):
-            raise TypeError('text argument must be str (%r)' % type(text))
-
-        if self.content_type == 'application/octet-stream':
-            self.content_type = 'text/plain'
-        if self.charset is None:
-            self.charset = 'utf-8'
-
-        self.body = text.encode(self.charset)
-
-    @asyncio.coroutine
-    def write_eof(self):
-        body = self._body
-        if body is not None:
-            self.write(body)
-        yield from super().write_eof()
 
 
 class WebSocketResponse(StreamResponse):
@@ -900,246 +300,6 @@ class WebSocketResponse(StreamResponse):
 
     def write(self, data):
         raise RuntimeError("Cannot call .write() for websocket")
-
-
-############################################################
-# HTTP Exceptions
-############################################################
-
-class HTTPException(Response, Exception):
-
-    # You should set in subclasses:
-    # status = 200
-
-    status_code = None
-
-    def __init__(self, *, headers=None, reason=None,
-                 body=None, text=None, content_type=None):
-        Response.__init__(self, status=self.status_code,
-                          headers=headers, reason=reason,
-                          body=body, text=text, content_type=content_type)
-        Exception.__init__(self, self.reason)
-        if self.body is None:
-            self.text = "{}: {}".format(self.status, self.reason)
-
-
-class HTTPError(HTTPException):
-    """Base class for exceptions with status codes in the 400s and 500s."""
-
-
-class HTTPRedirection(HTTPException):
-    """Base class for exceptions with status codes in the 300s."""
-
-
-class HTTPSuccessful(HTTPException):
-    """Base class for exceptions with status codes in the 200s."""
-
-
-class HTTPOk(HTTPSuccessful):
-    status_code = 200
-
-
-class HTTPCreated(HTTPSuccessful):
-    status_code = 201
-
-
-class HTTPAccepted(HTTPSuccessful):
-    status_code = 202
-
-
-class HTTPNonAuthoritativeInformation(HTTPSuccessful):
-    status_code = 203
-
-
-class HTTPNoContent(HTTPSuccessful):
-    status_code = 204
-
-
-class HTTPResetContent(HTTPSuccessful):
-    status_code = 205
-
-
-class HTTPPartialContent(HTTPSuccessful):
-    status_code = 206
-
-
-############################################################
-# 3xx redirection
-############################################################
-
-
-class _HTTPMove(HTTPRedirection):
-
-    def __init__(self, location, *, headers=None, reason=None,
-                 body=None, text=None, content_type=None):
-        if not location:
-            raise ValueError("HTTP redirects need a location to redirect to.")
-        super().__init__(headers=headers, reason=reason,
-                         body=body, text=text, content_type=content_type)
-        self.headers['Location'] = location
-        self.location = location
-
-
-class HTTPMultipleChoices(_HTTPMove):
-    status_code = 300
-
-
-class HTTPMovedPermanently(_HTTPMove):
-    status_code = 301
-
-
-class HTTPFound(_HTTPMove):
-    status_code = 302
-
-
-# This one is safe after a POST (the redirected location will be
-# retrieved with GET):
-class HTTPSeeOther(_HTTPMove):
-    status_code = 303
-
-
-class HTTPNotModified(HTTPRedirection):
-    # FIXME: this should include a date or etag header
-    status_code = 304
-
-
-class HTTPUseProxy(_HTTPMove):
-    # Not a move, but looks a little like one
-    status_code = 305
-
-
-class HTTPTemporaryRedirect(_HTTPMove):
-    status_code = 307
-
-
-############################################################
-# 4xx client error
-############################################################
-
-
-class HTTPClientError(HTTPError):
-    pass
-
-
-class HTTPBadRequest(HTTPClientError):
-    status_code = 400
-
-
-class HTTPUnauthorized(HTTPClientError):
-    status_code = 401
-
-
-class HTTPPaymentRequired(HTTPClientError):
-    status_code = 402
-
-
-class HTTPForbidden(HTTPClientError):
-    status_code = 403
-
-
-class HTTPNotFound(HTTPClientError):
-    status_code = 404
-
-
-class HTTPMethodNotAllowed(HTTPClientError):
-    status_code = 405
-
-    def __init__(self, method, allowed_methods, *, headers=None, reason=None,
-                 body=None, text=None, content_type=None):
-        allow = ','.join(sorted(allowed_methods))
-        super().__init__(headers=headers, reason=reason,
-                         body=body, text=text, content_type=content_type)
-        self.headers['Allow'] = allow
-        self.allowed_methods = allowed_methods
-        self.method = method.upper()
-
-
-class HTTPNotAcceptable(HTTPClientError):
-    status_code = 406
-
-
-class HTTPProxyAuthenticationRequired(HTTPClientError):
-    status_code = 407
-
-
-class HTTPRequestTimeout(HTTPClientError):
-    status_code = 408
-
-
-class HTTPConflict(HTTPClientError):
-    status_code = 409
-
-
-class HTTPGone(HTTPClientError):
-    status_code = 410
-
-
-class HTTPLengthRequired(HTTPClientError):
-    status_code = 411
-
-
-class HTTPPreconditionFailed(HTTPClientError):
-    status_code = 412
-
-
-class HTTPRequestEntityTooLarge(HTTPClientError):
-    status_code = 413
-
-
-class HTTPRequestURITooLong(HTTPClientError):
-    status_code = 414
-
-
-class HTTPUnsupportedMediaType(HTTPClientError):
-    status_code = 415
-
-
-class HTTPRequestRangeNotSatisfiable(HTTPClientError):
-    status_code = 416
-
-
-class HTTPExpectationFailed(HTTPClientError):
-    status_code = 417
-
-
-############################################################
-# 5xx Server Error
-############################################################
-#  Response status codes beginning with the digit "5" indicate cases in
-#  which the server is aware that it has erred or is incapable of
-#  performing the request. Except when responding to a HEAD request, the
-#  server SHOULD include an entity containing an explanation of the error
-#  situation, and whether it is a temporary or permanent condition. User
-#  agents SHOULD display any included entity to the user. These response
-#  codes are applicable to any request method.
-
-
-class HTTPServerError(HTTPError):
-    pass
-
-
-class HTTPInternalServerError(HTTPServerError):
-    status_code = 500
-
-
-class HTTPNotImplemented(HTTPServerError):
-    status_code = 501
-
-
-class HTTPBadGateway(HTTPServerError):
-    status_code = 502
-
-
-class HTTPServiceUnavailable(HTTPServerError):
-    status_code = 503
-
-
-class HTTPGatewayTimeout(HTTPServerError):
-    status_code = 504
-
-
-class HTTPVersionNotSupported(HTTPServerError):
-    status_code = 505
 
 
 ############################################################
